% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daylength.R
\name{DayLength}
\alias{DayLength}
\title{Day Length Calculation}
\usage{
DayLength(n, phi)
}
\arguments{
\item{n}{An integer representing the day of the year (1-365).}

\item{phi}{A numeric value representing the latitude in degrees (-90 to 90).}
}
\value{
A numeric value representing the length of the day in hours.
}
\description{
Calculate the length of the day in hours for a given day number and latitude.
}
\examples{
DayLength(80, 45) # calculates the day length in hours for March 21st at latitude 45 degrees

}
