% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelvin_to_celcius.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Convert temperature in Kelvin to Celsius}
\usage{
kelvin_to_celsius(temp_k)
}
\arguments{
\item{temp_k}{A temperature in Kelvin}
}
\value{
A temperature in Celsius
}
\description{
This function takes in a temperature in Kelvin and returns the
equivalent temperature in Celsius.
}
\examples{
kelvin_to_celsius(273.15)
kelvin_to_celsius(373.15)
}
