% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solarelev.R
\name{solarElev}
\alias{solarElev}
\title{Compute solar elevation angles for a year}
\usage{
solarElev(phi)
}
\arguments{
\item{phi}{A numeric value representing the latitude in degrees.}
}
\value{
A numeric vector of length 365 representing the solar elevation angles in degrees.
}
\description{
This function calculates the solar elevation angles for each day of a year for a given latitude.
}
\examples{
solarElev(49.01)

}
