% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardXYPlot.R
\name{StandardXYPlot}
\alias{StandardXYPlot}
\title{Plot Two Vectors}
\usage{
StandardXYPlot(x, y, title, xlab, ylab, linecolor, linenames)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{title}{The title of the plot.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{linecolor}{The color of the plotted line.}

\item{linenames}{A character vector of length 2 containing the names of the lines to be plotted.}
}
\value{
A plot of the two vectors with a legend.
}
\description{
This function plots two vectors against each other.
}
\examples{
StandardXYPlot(c(1, 2, 3), c(4, 5, 6), "My Plot", "X Axis", "Y Axis", "blue", c("Line 1", "Line 2"))

}
