% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sundown.R
\name{Sundown}
\alias{Sundown}
\title{Calculate the time of sundown for a given day and location}
\usage{
Sundown(n, phi)
}
\arguments{
\item{n}{An integer representing the day of the year (1 = January 1, 365 or 366 = December 31)}

\item{phi}{A numeric value representing the latitude of the location in degrees}
}
\value{
A numeric value representing the time of sundown in hours (e.g. 18.75 = 6:45 PM)
}
\description{
This function calculates the time of sundown for a given day and location, based on the method described in "Astronomical Algorithms" by Jean Meeus.
}
\examples{
Sundown(200, 52.5) # calculates the time of sundown for July 19th at 52.5 degrees north latitude

}
